move status handling to status code
