+++ lib/charms/mongodb/v1/helpers.py
@@ -10,15 +10,13 @@ import string
 import subprocess
 from typing import List
 
-from charms.mongodb.v0.mongodb import MongoDBConfiguration, MongoDBConnection
+from charms.mongodb.v0.mongodb import MongoDBConfiguration
 from ops.model import (
     ActiveStatus,
-    BlockedStatus,
     MaintenanceStatus,
     StatusBase,
     WaitingStatus,
 )
-from pymongo.errors import AutoReconnect, ServerSelectionTimeoutError
 
 from config import Config
 
@@ -271,38 +269,6 @@ def generate_keyfile() -> str:
     """
     choices = string.ascii_letters + string.digits
     return "".join([secrets.choice(choices) for _ in range(1024)])
-
-
-def build_unit_status(mongodb_config: MongoDBConfiguration, unit_ip: str) -> StatusBase:
-    """Generates the status of a unit based on its status reported by mongod."""
-    try:
-        with MongoDBConnection(mongodb_config) as mongo:
-            replset_status = mongo.get_replset_status()
-
-            if unit_ip not in replset_status:
-                return WaitingStatus("Member being added..")
-
-            replica_status = replset_status[unit_ip]
-
-            if replica_status == "PRIMARY":
-                return ActiveStatus("Primary")
-            elif replica_status == "SECONDARY":
-                return ActiveStatus("")
-            elif replica_status in ["STARTUP", "STARTUP2", "ROLLBACK", "RECOVERING"]:
-                return WaitingStatus("Member is syncing...")
-            elif replica_status == "REMOVED":
-                return WaitingStatus("Member is removing...")
-            else:
-                return BlockedStatus(replica_status)
-    except ServerSelectionTimeoutError as e:
-        # ServerSelectionTimeoutError is commonly due to ReplicaSetNoPrimary
-        logger.debug("Got error: %s, while checking replica set status", str(e))
-        return WaitingStatus("Waiting for primary re-election..")
-    except AutoReconnect as e:
-        # AutoReconnect is raised when a connection to the database is lost and an attempt to
-        # auto-reconnect will be made by pymongo.
-        logger.debug("Got error: %s, while checking replica set status", str(e))
-        return WaitingStatus("Waiting to reconnect to unit..")
 
 
 def copy_licenses_to_unit():
